Extendbenchmark

+++ src/de/upb/crc901/proseco/prototype/genderpredictor/benchmark/Benchmark.java
@@ -1,19 +1,19 @@
 package de.upb.crc901.proseco.prototype.genderpredictor.benchmark;
 
 import java.io.BufferedReader;
-import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.lang.ProcessBuilder.Redirect;
+import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Set;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
+import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import org.apache.commons.io.FileUtils;
@@ -21,7 +21,6 @@ import org.apache.commons.io.FilenameUtils;
 
 import de.upb.crc901.proseco.PrototypeProperties;
 import de.upb.crc901.proseco.prototype.genderpredictor.GroundingRoutine;
-import jaicore.basic.FileUtil;
 import jaicore.basic.PerformanceLogger;
 
 public class Benchmark extends Thread {
@@ -32,10 +31,6 @@ public class Benchmark extends Thread {
 	private static final File FINISHED_TASK_DIR = new File(PROPS.getProperty("finished_task_dir"));
 	private static final File TESTBED_DIR = new File(PROPS.getProperty("testbed_dir"));
 
-	private static final String FVALUE_FILE = PROPS.getProperty("name_fvaluefile");
-	private static final String CLASSIFIER_MODEL_FILE = PROPS.getProperty("classifier_model_file");
-
-	private static final File BENCHMARK_INSTANCES_FILE = new File(PROPS.getProperty("benchmark_instances"));
 	private static final File SOURCE_INPUT_FOLDER = new File(PROPS.getProperty("prototype_source_code"));
 
 	private static final int NUMBER_OF_THREADS = Integer.parseInt(PROPS.getProperty("number_of_threads"));
@@ -71,88 +66,49 @@ public class Benchmark extends Thread {
 						continue;
 					}
 
+					BenchmarkTask task;
 					this.fileNameListLock.lock();
 					try {
 						if(this.processedFileNames.contains(taskFile.getName())) {
 							continue;
 						} else {
 							this.processedFileNames.add(taskFile.getName());
+							task = BenchmarkTask.readFromTaskFile(taskFile);
+							this.taskTempFolder = new File(TESTBED_DIR.getAbsoluteFile() + "_" + task.getCandidateFolder().getName());
+							FileUtils.copyDirectory(TESTBED_DIR, this.taskTempFolder);
 						}
-					}finally {
+					} finally {
 						this.fileNameListLock.unlock();
 					}
 
-					final String taskOutputFolder = FileUtil.readFileAsString(taskFile.getAbsolutePath()).trim();
-					if (taskOutputFolder.equals("")) {
-						continue;
+
+					GroundingRoutine groundingRoutine = new GroundingRoutine(task.getCandidateFolder(), SOURCE_INPUT_FOLDER.getCanonicalFile(), this.taskTempFolder);
+
+					AbstractBenchmarkRunner benchmarkRunner = null;
+					switch(task.getBuildPhase()) {
+					case FEATURE_EXTRACTION:
+						benchmarkRunner = new FeatureExtractionBenchmarkRunner(task, groundingRoutine, this.taskTempFolder);
+						break;
+					case CLASSIFIER_DEF:
+						benchmarkRunner = new ClassifierBenchmarkRunner(task, groundingRoutine, this.taskTempFolder);
+						break;
 					}
 
-					final File candidateFolder = new File(taskOutputFolder);
+					log("Start to benchmark task " + taskFile.getAbsolutePath() + " for candidate " + task.getCandidateFolder().getAbsolutePath());
+					benchmarkRunner.run();
 
-					long totalBenchmarkStartTime = System.currentTimeMillis();
-
-					this.taskTempFolder = new File(TESTBED_DIR.getAbsoluteFile() + "_" + candidateFolder.getName());
-					FileUtils.copyDirectory(TESTBED_DIR, this.taskTempFolder);
-
-
-					try (BufferedWriter bw = new BufferedWriter(new FileWriter(this.taskTempFolder.getAbsolutePath() + File.separator + "runtime.value"))) {
-						bw.write(candidateFolder.getName()+"\n");
-					} catch(IOException ioE) {
-						ioE.printStackTrace();
-					}
-					long startTime;
-
-					PerformanceLogger.logStart("PerformBenchmarkForCandidate");
-					log("Start to benchmark task " + taskFile.getAbsolutePath() + " for candidate " + candidateFolder.getAbsolutePath());
-
-					// Execute grounding => code assembly + compile + training
-					log("Execute Grounding routine...");
-					startTime = System.currentTimeMillis();
-					final String[] groundingParams = { candidateFolder.getAbsolutePath(), SOURCE_INPUT_FOLDER.getCanonicalPath(), this.taskTempFolder.getAbsolutePath() };
-					GroundingRoutine.main(groundingParams);
-
-					try (BufferedWriter bw = new BufferedWriter(new FileWriter(this.taskTempFolder.getAbsolutePath() + File.separator + "runtime.value",true))) {
-						bw.write("totalGrounding="+(System.currentTimeMillis()-startTime)+"ms\n");
-					} catch(IOException ioE) {
-						ioE.printStackTrace();
-					}
-					log("Grounding routine finished.");
-
-					// Test trained instance against validation set
-					PerformanceLogger.logStart("computeFValue");
-					log("Compute f value for current testbed...", false);
-					startTime = System.currentTimeMillis();
-					this.computeFValue();
-					log("DONE");
-					PerformanceLogger.logEnd("computeFValue");
-					try (BufferedWriter bw = new BufferedWriter(new FileWriter(this.taskTempFolder.getAbsolutePath() + File.separator + "runtime.value",true))) {
-						bw.write("fValueComputation="+(System.currentTimeMillis()-startTime)+"ms\n");
-					} catch(IOException ioE) {
-						ioE.printStackTrace();
-					}
-
-					try (BufferedWriter bw = new BufferedWriter(new FileWriter(this.taskTempFolder.getAbsolutePath() + File.separator + "runtime.value",true))) {
-						bw.write("totalBenchmark="+(System.currentTimeMillis()-totalBenchmarkStartTime)+"ms\n");
-					} catch(IOException ioE) {
-						ioE.printStackTrace();
-					}
+					String[] ignoreFilesForMoving = {"compile.bat", "libs", "test.bat", "train.bat", "validationInstances.serialized"};
+					Set<String> ignoreFilesForMovingSet = Arrays.stream(ignoreFilesForMoving).collect(Collectors.toSet());
 
 					// move task specific files to task directory
 					log("Benchmark Service: Move files...", false);
 					for (final File testBedFile : this.taskTempFolder.listFiles()) {
-						switch (testBedFile.getName()) {
-						case "classifier.model":
-						case "GenderPredictor.java":
-						case "GenderPredictor.class":
-						case "f.value":
-						case "runtime.value":
-							final File candidateFile = new File(candidateFolder.getAbsolutePath() + File.separator + testBedFile.getName());
-
+						if (!ignoreFilesForMovingSet.contains(testBedFile.getName())) {
+							final File candidateFile = new File(task.getCandidateFolder().getAbsolutePath() + File.separator + testBedFile.getName());
 							if (candidateFile.exists()) {
 								candidateFile.delete();
 							}
 							FileUtils.copyFile(testBedFile, candidateFile);
-							break;
 						}
 					}
 					FileUtils.copyFile(taskFile, new File(FINISHED_TASK_DIR.getAbsolutePath() + File.separator + taskFile.getName()));
@@ -185,41 +141,6 @@ public class Benchmark extends Thread {
 		this.keepRunning = false;
 	}
 
-	private void computeFValue() {
-		double f = 0;
-		if (new File(this.taskTempFolder + File.separator + CLASSIFIER_MODEL_FILE).exists()) {
-			final ProcessBuilder pb = new ProcessBuilder(this.taskTempFolder.getAbsolutePath() + File.separator + "test.bat", BENCHMARK_INSTANCES_FILE.getAbsolutePath());
-			pb.redirectError(Redirect.INHERIT);
-
-			try {
-				final Process fValueProcess = pb.start();
-				try (BufferedReader br = new BufferedReader(new InputStreamReader(fValueProcess.getInputStream()), 1)) {
-					final String fValueString = br.readLine();
-					if (fValueString != null) {
-
-						final String[] fValueStringSplit = fValueString.split("=");
-						if (fValueStringSplit.length == 2) {
-							f = Double.parseDouble(fValueStringSplit[1]);
-						}
-					}
-				}
-				fValueProcess.waitFor();
-			} catch (final IOException e) {
-				e.printStackTrace();
-			} catch (final InterruptedException e) {
-				e.printStackTrace();
-			}
-		}
-
-		try (BufferedWriter bw = new BufferedWriter(new FileWriter(this.taskTempFolder + File.separator + FVALUE_FILE))) {
-			bw.write(f + "\n");
-		} catch (final IOException e) {
-			e.printStackTrace();
-		}
-	}
-
-	private static boolean lastLineBreak = true;
-
 	public static void main(final String[] args) {
 		ExecutorService threadPool = Executors.newFixedThreadPool(NUMBER_OF_THREADS);
 		List<String> taskFilenameList = new LinkedList<>();
@@ -233,7 +154,6 @@ public class Benchmark extends Thread {
 			while ((line = br.readLine()) != null) {
 				switch (line.trim()) {
 				case "q":
-					System.err.println("Blaa");
 					threadPool.shutdownNow();
 					threadPool.awaitTermination(2000, TimeUnit.MILLISECONDS);
 					System.exit(0);
@@ -251,6 +171,7 @@ public class Benchmark extends Thread {
 		log(msg, true);
 	}
 
+	private static boolean lastLineBreak = true;
 	private static void log(final String msg, final boolean linebreak) {
 		final String prefix = "[" + Thread.currentThread().getName() + "] Benchmark Service: ";
 		String printString;
